Add weekly digest and feed endpoints, validation, and score function
